+++ shuup/core/models/_contacts.py
@@ -20,6 +20,10 @@ from timezone_field.fields import TimeZoneField
 from shuup import configuration
 from shuup.core.fields import InternalIdentifierField, LanguageField
 from shuup.core.pricing import PriceDisplayOptions
+from shuup.core.utils.users import (
+    is_user_all_seeing, should_force_anonymous_contact,
+    should_force_person_contact
+)
 from shuup.utils.analog import define_log_model
 
 from ._base import PolymorphicShuupModel, TranslatableShuupModel
@@ -354,10 +358,7 @@ class PersonContact(Contact):
 
     @cached_property
     def is_all_seeing(self):
-        if self.user_id and self.user.is_superuser:
-            all_seeing_key = "is_all_seeing:%d" % self.user_id
-            return configuration.get(None, all_seeing_key, False)
-        return False
+        return is_user_all_seeing(self.user)
 
 
 class AnonymousContact(Contact):
@@ -432,6 +433,10 @@ def get_person_contact(user):
     """
     if not user or user.is_anonymous():
         return AnonymousContact()
+
+    if should_force_anonymous_contact(user):
+        return AnonymousContact()
+
     defaults = {
         'is_active': user.is_active,
         'first_name': getattr(user, 'first_name', ''),
@@ -454,10 +459,31 @@ def get_company_contact(user):
       CompanyContact (or none) of which user's PersonContact is a member
     :rtype: CompanyContact|None
     """
+    if should_force_person_contact(user):
+        return None
+
     contact = get_person_contact(user)
     if not contact:
         return None
     return contact.company_memberships.filter(is_active=True).first()
+
+
+def get_company_contact_for_shop(shop):
+    company = CompanyContact.objects.get_or_create(
+        identifier="shop-contact-%s" % shop.pk, defaults={"name": shop.public_name}
+    )[0]
+    company.shops.add(shop)
+    return company
+
+
+def get_company_contact_for_shop_staff(shop, user):
+    assert user.is_staff
+    if not getattr(user, "is_superuser", False):
+        assert shop.staff_members.filter(id=user.id).exists()
+
+    contact = get_company_contact_for_shop(shop)
+    contact.members.add(get_person_contact(user))
+    return contact
 
 
 CompanyContactLogEntry = define_log_model(CompanyContact)